Add score, player information, and new game button
These items required adjusting some other code to make the board
remain responsive, centered, and a decent size. The footer was
removed, because the trick used to make the board responsive causes
some space to be below it, and that makes the footer go to the next
screen on phones, even though there is visually plenty of space for
it. The page looks better without it.